frontend_streamlit/requirements.txt
streamlit==1.36.0
streamlit-drawable-canvas==0.9.3
pymupdf==1.24.2
Pillow==10.3.0
PyYAML==6.0.1
